+++ scripts/utils/__init__.py
@@ -1,2 +1,4 @@
 from . import metadata_interaction
 from . import network_interaction
+from . import chain_model
+from . import getting_data
