+++ scripts/utils/chain_model.py
@@ -1,5 +1,5 @@
-from scripts.utils.network_interaction import create_connection_by_url
-from scripts.create_type_file import get_properties
+from utils.network_interaction import create_connection_by_url
+from .metadata_interaction import get_properties
 
 
 class Chain():
@@ -27,7 +27,7 @@ class Chain():
                 print("Can't connect to that node")
                 continue
 
-        raise TimeoutError("Can't connect to all nodes of network", self.name)
+        print("Can't connect to all nodes of network", self.name)
 
 
     def init_properties(self):
