+++ tests/data/setting_data.py
@@ -1,6 +1,6 @@
 import os
-from tests.utils.getting_data import get_network_list
-from tests.utils.chain_model import Chain
+from scripts.utils.getting_data import get_network_list
+from scripts.utils.chain_model import Chain
 
 network_file_path = os.getenv('CHAINS_JSON_PATH', "chains/v6/chains.json")
 skipped_networks = ['Edgeware']
