+++ tests/test_can_query_base_parameters.py
@@ -1,5 +1,5 @@
 import pytest
-from tests.utils.chain_model import Chain
+from scripts.utils.chain_model import Chain
 from substrateinterface import SubstrateInterface
 from tests.data.setting_data import *
