+++ tests/test_nodes_availability.py
@@ -4,7 +4,7 @@ import pytest
 
 from tests.data.setting_data import *
 from scripts.utils.network_interaction import create_connection_by_url
-from tests.utils.chain_model import Chain
+from scripts.utils.chain_model import Chain
 from substrateinterface import SubstrateInterface
