+++ tests/test_xcm_data.py
@@ -3,7 +3,7 @@ import json
 import deepdiff
 
 from tests.data.setting_data import *
-from tests.utils.chain_model import Chain
+from scripts.utils.chain_model import Chain
 
 xcm_data_file_path = './tests/data/xcm_data.json'
